requirements.txt
streamlit
sqlalchemy
alembic
python-dotenv
google-generativeai
graphviz
cognee-core
lancedb